requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-crispy-forms==1.11.0
pytz==2021.1
soupsieve==2.2
sqlparse==0.4.1
